move duplicated code to an extension method.
